[not verified] Move types file
