+++ projects/plugins/crm/src/js/state/automations-admin/types.ts
